Added help command & fixed formatting errors
- Added help command, which was missing
- Fixed formatting error with many commands (double curly brackets were used instead of single ones, eg {{}} instead of {}

It should also be noted that there is an issue with cv2 and compiling to an exe (on py3.9 it seems), i put a comment explaining whats wrong. To fix it, you would need to add --paths to the pyinstaller command, which would point to Python39\site-packages\cv2\python-3.9 . I have not included this fix in this pull request.