Hidden Window For RAT
The RAT opens a window when it is run, making it impractical for general use. My edit simply adds an option to hide the window.
Unfortunately, when I was compiling the RAT for myself I ran into a problem where when I ran the final result it would give an error with something along the lines of "ImportError: OpenCV loader: missing configuration file: ['config.py']. Check OpenCV installation." However, I was able to fix this by downgrading opencv to version 4.5.3.56 with "pip install opencv-python==4.5.3.56" This runs the RAT without a window (obviously it will be shown in task manager, there is a solution to that but it is way more advanced.)
I really like the work that you've done with this program, and can't wait to see more!
(Yes, I really just added 3 characters and think it's good enough for submission XD)

All local testing and workarounds made on Windows 10. I have not tested, nor do I plan to, on Linux. From my research, Linux users run into the same problem with PyInstaller on Ubuntu.